refactor: use css modules
- Refactor Select
- Refactor deposit form
