+++ src/features/deposit/components/Form/DepositFormNetwork/styles.module.css
@@ -1,0 +1,53 @@
+.container {
+  width: 100%;
+}
+
+@media (min-width: 768px) {
+  .container {
+    max-width: 472px;
+  }
+}
+
+.formWrapper {
+  border-radius: 1rem;
+  padding: 1.25rem;
+  background-color: white;
+  box-shadow: 0 1px 3px 0 rgb(0 0 0 / 0.1), 0 1px 2px -1px rgb(0 0 0 / 0.1);
+}
+
+[data-theme='dark'] .formWrapper {
+  background-color: rgba(17, 17, 16, 1);
+  box-shadow: 0 1px 3px 0 rgb(255 255 255 / 0.1), 0 1px 2px -1px rgb(255 255 255 / 0.1);
+}
+
+.selectWrapper {
+  margin-bottom: 1rem;
+}
+
+.buttonGroup {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.button {
+  width: 100%;
+  height: 56px;
+  border-radius: 0.5rem;
+}
+
+.buttonContent {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.orangeButton {
+  background-color: orange;
+  color: white;
+}
+
+.orangeButton:hover {
+  background-color: darkorange;
+}
