+++ src/types/css-modules.d.ts
@@ -1,0 +1,4 @@
+declare module "*.module.css" {
+  const classes: { [key: string]: string }
+  export default classes
+}
